Correzzione tipo variabili ID per il DB

+++ ProjectWork/src/main/java/com/example/ProjectWork/model/Candidato.java
@@ -2,7 +2,6 @@ package com.example.ProjectWork.model;
 
 import jakarta.persistence.*;
 
-import java.math.BigInteger;
 
 @Entity
 @Table(name = "CANDIDATO", schema = "dbo")
@@ -10,7 +9,7 @@ public class Candidato {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer idCandidato;
+    private Long idCandidato;
 
     @OneToOne
     @JoinColumn(name = "idUtente")
